pyproject.toml
[tool.poetry]
name = "local-intelligence-hub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">3.11,<3.13"
Django = "4.2.11"
django-bootstrap5 = "^22.1"
django-environ = "^0.9.0"
psycopg2 = "^2.9.3"
django-libsass = "^0.9"
requests-cache = "^0.9.6"
Pillow = "^10.2.0"
python-magic = "^0.4.27"
tqdm = "^4.64.1"
openpyxl = "^3.0.10"
mysoc-dataset = "^0.3.0"
django-jsonform = "^2.15.0"
lxml = "^4.9.2"
beautifulsoup4 = "^4.11.1"
django-polymorphic = "^3.1.0"
pyairtable = "^2.3.0.post1"
procrastinate = {version = "2.0.0b4", extras = ["django"]}
psycopg = "^3.1.18"
aiopg = "^1.4.0"
strawberry-graphql-django = "^0.32.1"
strawberry-django-auth = "^0.376.5"
django-cors-headers = "^4.3.1"
python-benedict = "^0.33.2"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
setuptools = "^69.1.1"
uvicorn = "^0.27.1"
strawberry-graphql = {extras = ["asgi", "opentelemetry"], version = "^0.226.0"}
pandas = "^2.2.1"
django-choices-field = "^2.3.0"
httpx = "^0.27.0"
strawberry-django-dataloaders = "^0.2.0"
sentry-sdk = {extras = ["opentelemetry"], version = "^1.45.0"}
django-minio-storage = "^0.5.7"
mailchimp3 = "^3.0.21"
scalene = "^1.5.38"
django-vectortiles = { git = "https://github.com/submarcos/django-vectortiles.git" }
django-cryptography = "^1.1"
posthog = "^3.5.0"
mailchimp-marketing = {git = "https://github.com/mailchimp/mailchimp-marketing-python.git"}
cryptography = "^42.0.5"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^4.3"
black = "^24.3.0"
coverage = "^6.5.0"
flake8 = "^6.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
default_section = "THIRDPARTY"
known_first_party = "hub"
known_django = "django"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
extend_skip = ["migrations"]
